user

+++ src/main/java/com/yimei/finance/entity/admin/finance/FinanceOrder.java
@@ -25,7 +25,7 @@ public class FinanceOrder extends BaseEntity implements Serializable {
     private String sourceId;                                         //流水号，编号
 
     @Column(name = "user_id")
-    private int userId;                                              //申请人用户id
+    private Long userId;                                              //申请人用户id
 
     @Column(name = "apply_user_name")
     private String applyUserName;                                    //申请人姓名
